add local cart logic

+++ src/client/actions/cart.js
@@ -2,6 +2,7 @@ import axios from "axios";
 
 import { addAlert } from "./alert";
 import { AlertTypes } from "../Constants";
+import store from "../store";
 
 import {
   GET_CART,
@@ -13,9 +14,46 @@ import {
   APPLY_COUPON,
   REMOVE_COUPON,
   CLEAR_CART_ERRORS,
+  GET_LOCAL_CART,
+  SYNC_LOCAL_CART,
 } from "./types";
 
+const handleErrors =
+  (err, isToAddToAlert = false, isToAddToCartState = false) =>
+  async (dispatch) => {
+    console.error(err);
+
+    if (err.response.data.errors) {
+      const errors = err.response.data.errors;
+
+      if (isToAddToAlert) {
+        errors.forEach((error) =>
+          dispatch(addAlert(error.msg, AlertTypes.INFO))
+        );
+      }
+
+      if (isToAddToCartState) {
+        dispatch({
+          type: CART_ERROR,
+          payload: err.response.data.errors
+            .map((error) => error.msg)
+            .join(","),
+        });
+
+        setTimeout(() => {
+          clearCartErrors()(dispatch);
+        }, 2000);
+      }
+    }
+  };
+
 export const getCart = () => async (dispatch) => {
+  const isAuthenticated = store.getState().authReducer.isAuthenticated;
+  if (!isAuthenticated) {
+    getLocalCart()(dispatch);
+    return;
+  }
+
   try {
     const res = await axios.get("/api/cart");
 
@@ -24,26 +62,17 @@ export const getCart = () => async (dispatch) => {
       payload: res.data,
     });
   } catch (err) {
-    console.error(err.response.data);
-
-    const errors = err.response.data.errors;
-
-    if (errors) {
-      dispatch({
-        type: CART_ERROR,
-        payload: err.response.data.errors
-          .map((error) => error.msg)
-          .join(","),
-      });
-
-      errors.forEach((error) =>
-        dispatch(addAlert(error.msg, AlertTypes.INFO))
-      );
-    }
+    handleErrors(err, true, false)(dispatch);
   }
 };
 
 export const addToCart = (productId) => async (dispatch) => {
+  const isAuthenticated = store.getState().authReducer.isAuthenticated;
+  if (!isAuthenticated) {
+    dispatch(addToLocalCart(productId));
+    return;
+  }
+
   try {
     const res = await axios.post(`/api/cart/`, { productId });
 
@@ -51,27 +80,19 @@ export const addToCart = (productId) => async (dispatch) => {
       type: ADD_TO_CART,
       payload: res.data,
     });
+
+    dispatch(getCart());
   } catch (err) {
-    console.error(err.response.data);
-
-    const errors = err.response.data.errors;
-
-    if (errors) {
-      dispatch({
-        type: CART_ERROR,
-        payload: err.response.data.errors
-          .map((error) => error.msg)
-          .join(","),
-      });
-
-      errors.forEach((error) =>
-        dispatch(addAlert(error.msg, AlertTypes.INFO))
-      );
-    }
+    handleErrors(err, true, false)(dispatch);
   }
 };
 
 export const removeOneFromCart = (productId) => async (dispatch) => {
+  const isAuthenticated = store.getState().authReducer.isAuthenticated;
+  if (!isAuthenticated) {
+    dispatch(removeOneFromLocalCart(productId));
+    return;
+  }
   try {
     const res = await axios.put(`/api/cart/`, { productId });
 
@@ -79,26 +100,21 @@ export const removeOneFromCart = (productId) => async (dispatch) => {
       type: REMOVE_ONE_FROM_CART,
       payload: res.data,
     });
+
+    dispatch(getCart());
   } catch (err) {
-    console.error(err.response.data);
-
-    const errors = err.response.data.errors;
-    if (errors) {
-      dispatch({
-        type: CART_ERROR,
-        payload: err.response.data.errors
-          .map((error) => error.msg)
-          .join(","),
-      });
-
-      errors.forEach((error) =>
-        dispatch(addAlert(error.msg, AlertTypes.INFO))
-      );
-    }
+    handleErrors(err, true, false)(dispatch);
   }
 };
 
 export const removeAllFromCart = (productId) => async (dispatch) => {
+  const isAuthenticated = store.getState().authReducer.isAuthenticated;
+  if (!isAuthenticated) {
+    console.log("removeAllFromLocalCart");
+    removeAllFromLocalCart(productId)(dispatch);
+    return;
+  }
+
   try {
     // delete do not have a header, so need to use data to add header
     const res = await axios.delete(`/api/cart/`, { data: { productId } });
@@ -107,22 +123,10 @@ export const removeAllFromCart = (productId) => async (dispatch) => {
       type: REMOVE_ALL_FROM_CART,
       payload: res.data,
     });
+
+    dispatch(getCart());
   } catch (err) {
-    console.error(err.response.data);
-
-    const errors = err.response.data.errors;
-    if (errors) {
-      dispatch({
-        type: CART_ERROR,
-        payload: err.response.data.errors
-          .map((error) => error.msg)
-          .join(","),
-      });
-
-      errors.forEach((error) =>
-        dispatch(addAlert(error.msg, AlertTypes.INFO))
-      );
-    }
+    handleErrors(err, true, false)(dispatch);
   }
 };
 
@@ -130,6 +134,8 @@ export const clearLocalCart = () => (dispatch) => {
   dispatch({
     type: CLEAR_LOCAL_CART,
   });
+
+  dispatch(getCart());
 };
 
 export const applyCoupon = (coupon) => async (dispatch) => {
@@ -140,26 +146,10 @@ export const applyCoupon = (coupon) => async (dispatch) => {
       type: APPLY_COUPON,
       payload: res.data,
     });
+
+    dispatch(getCart());
   } catch (err) {
-    console.error(err.response.data);
-
-    const errors = err.response.data.errors;
-    if (errors) {
-      dispatch({
-        type: CART_ERROR,
-        payload: err.response.data.errors
-          .map((error) => error.msg)
-          .join(","),
-      });
-
-      // errors.forEach((error) =>
-      //   dispatch(addAlert(error.msg, AlertTypes.INFO))
-      // );
-
-      setTimeout(() => {
-        dispatch(clearCartErrors());
-      }, 2000);
-    }
+    handleErrors(err, true, true)(dispatch);
   }
 };
 
@@ -173,26 +163,10 @@ export const removeCoupon = (couponId) => async (dispatch) => {
       type: REMOVE_COUPON,
       payload: res.data,
     });
+
+    dispatch(getCart());
   } catch (err) {
-    console.error(err.response.data);
-
-    const errors = err.response.data.errors;
-    if (errors) {
-      dispatch({
-        type: CART_ERROR,
-        payload: err.response.data.errors
-          .map((error) => error.msg)
-          .join(","),
-      });
-
-      // errors.forEach((error) =>
-      //   dispatch(addAlert(error.msg, AlertTypes.INFO))
-      // );
-
-      setTimeout(() => {
-        dispatch(clearCartErrors());
-      }, 2000);
-    }
+    handleErrors(err, true, true)(dispatch);
   }
 };
 
@@ -200,4 +174,168 @@ export const clearCartErrors = () => (dispatch) => {
   dispatch({
     type: CLEAR_CART_ERRORS,
   });
+
+  dispatch(getCart());
+};
+
+export const getLocalCart = () => (dispatch) => {
+  dispatch({
+    type: GET_LOCAL_CART,
+  });
+};
+
+const numberfy = (num) => Number(Number(num).toFixed(2));
+
+export const addToLocalCart = (productId) => async (dispatch) => {
+  try {
+    const res = await axios.get(`/api/products/${productId}`);
+    let product = res.data;
+
+    let cartState = store.getState().cartReducer;
+
+    const isProductInCart = cartState.products.some(
+      (p) => p._id === productId
+    );
+
+    if (isProductInCart) {
+      cartState.products = cartState.products.map((p) => {
+        if (p._id === productId) {
+          return {
+            ...p,
+            inCartQuantity: p.inCartQuantity + 1,
+          };
+        }
+        return p;
+      });
+    } else {
+      const { _id, name, price, image_url } = product;
+      product = { _id, name, price, image_url, inCartQuantity: 1 };
+
+      cartState.products.push(product);
+    }
+
+    cartState.totalQuantity += 1;
+    cartState.subtotal = numberfy(cartState.subtotal + product.price);
+    cartState.tax = numberfy(cartState.subtotal * 0.1);
+    cartState.total = Math.max(
+      0,
+      numberfy(cartState.subtotal + cartState.tax - cartState.discount)
+    );
+
+    dispatch({
+      type: ADD_TO_CART,
+      payload: cartState,
+    });
+
+    dispatch(getCart());
+  } catch (err) {
+    handleErrors(err, true, false)(dispatch);
+  }
+};
+
+export const removeOneFromLocalCart = (productId) => async (dispatch) => {
+  try {
+    let cartState = store.getState().cartReducer;
+
+    const productInCart = cartState.products.find(
+      (p) => p._id === productId
+    );
+
+    if (!productInCart) {
+      console.error("product not in cart");
+      return;
+    }
+
+    if (productInCart.inCartQuantity === 1) {
+      cartState.products = [
+        ...cartState.products.filter((p) => p._id !== productId),
+      ];
+    } else {
+      cartState.products = cartState.products.map((p) => {
+        if (p._id === productId) {
+          return {
+            ...p,
+            inCartQuantity: p.inCartQuantity - 1,
+          };
+        }
+        return p;
+      });
+    }
+
+    cartState.totalQuantity -= 1;
+    cartState.subtotal = numberfy(
+      cartState.subtotal - productInCart.price
+    );
+    cartState.tax = numberfy(cartState.subtotal * 0.1);
+    cartState.total = Math.max(
+      0,
+      numberfy(cartState.subtotal + cartState.tax - cartState.discount)
+    );
+
+    dispatch({
+      type: REMOVE_ONE_FROM_CART,
+      payload: cartState,
+    });
+
+    dispatch(getCart());
+  } catch (err) {
+    handleErrors(err, true, false)(dispatch);
+  }
+};
+
+export const removeAllFromLocalCart = (productId) => async (dispatch) => {
+  try {
+    let cartState = store.getState().cartReducer;
+
+    const productInCart = cartState.products.find(
+      (p) => p._id === productId
+    );
+
+    if (!productInCart) {
+      console.error("product not in cart");
+      return;
+    }
+
+    cartState.products = [
+      ...cartState.products.filter((p) => p._id !== productId),
+    ];
+
+    cartState.totalQuantity -= productInCart.inCartQuantity;
+    cartState.subtotal = numberfy(
+      cartState.subtotal -
+        productInCart.price * productInCart.inCartQuantity
+    );
+    cartState.tax = numberfy(cartState.subtotal * 0.1);
+    cartState.total = Math.max(
+      0,
+      numberfy(cartState.subtotal + cartState.tax - cartState.discount)
+    );
+
+    dispatch({
+      type: REMOVE_ALL_FROM_CART,
+      payload: cartState,
+    });
+
+    dispatch(getCart());
+  } catch (err) {
+    handleErrors(err, true, false)(dispatch);
+  }
+};
+
+export const syncLocalCart = () => async (dispatch) => {
+  try {
+    let cartState = store.getState().cartReducer;
+    if (!cartState.products.length) return;
+
+    const res = await axios.post(`/api/cart/sync`, {
+      products: cartState.products,
+    });
+
+    dispatch({
+      type: SYNC_LOCAL_CART,
+      payload: res.data,
+    });
+  } catch (err) {
+    handleErrors(err, true, false)(dispatch);
+  }
 };
